EL-1308 - merge case details and applicant details sections into non finance section
